Splits NativeUtils into multiple files, and cleans up include orders. Plus adds some documentation
